modify start server for index.ts.
